[Breadcrumbs] Move to the core
