+++ .size-limit.js
@@ -28,7 +28,7 @@ module.exports = [
     name: 'The size of the @material-ui/core modules',
     webpack: true,
     path: 'packages/material-ui/build/index.js',
-    limit: '91.2 KB',
+    limit: '92.2 KB',
   },
   {
     name: 'The size of the @material-ui/styles modules',
