+++ docs/src/pages/demos/breadcrumbs/CustomSeparator.js
@@ -2,9 +2,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import Paper from '@material-ui/core/Paper';
+import Breadcrumbs from '@material-ui/core/Breadcrumbs';
 import Typography from '@material-ui/core/Typography';
-import Breadcrumbs from '@material-ui/lab/Breadcrumbs';
 import Link from '@material-ui/core/Link';
+import NavigateNextIcon from '@material-ui/icons/NavigateNext';
 
 const styles = theme => ({
   root: {
@@ -21,12 +22,13 @@ function handleClick(event) {
   alert('You clicked a breadcrumb.'); // eslint-disable-line no-alert
 }
 
-function SimpleBreadcrumbs(props) {
+function CustomSeparator(props) {
   const { classes } = props;
+
   return (
     <div className={classes.root}>
       <Paper className={classes.paper}>
-        <Breadcrumbs arial-label="Breadcrumb">
+        <Breadcrumbs separator="›" arial-label="Breadcrumb">
           <Link color="inherit" href="/" onClick={handleClick}>
             Material-UI
           </Link>
@@ -38,29 +40,34 @@ function SimpleBreadcrumbs(props) {
       </Paper>
       <br />
       <Paper className={classes.paper}>
-        <Breadcrumbs arial-label="Breadcrumb">
+        <Breadcrumbs separator="-" arial-label="Breadcrumb">
           <Link color="inherit" href="/" onClick={handleClick}>
             Material-UI
           </Link>
           <Link color="inherit" href="/lab/about/" onClick={handleClick}>
             Lab
           </Link>
-          <Link
-            color="textPrimary"
-            href="/lab/about/breadcrumbs"
-            onClick={handleClick}
-            aria-current="page"
-          >
-            Breadcrumb
+          <Typography color="textPrimary">Breadcrumb</Typography>
+        </Breadcrumbs>
+      </Paper>
+      <br />
+      <Paper className={classes.paper}>
+        <Breadcrumbs separator={<NavigateNextIcon fontSize="small" />} arial-label="Breadcrumb">
+          <Link color="inherit" href="/" onClick={handleClick}>
+            Material-UI
           </Link>
+          <Link color="inherit" href="/lab/about/" onClick={handleClick}>
+            Lab
+          </Link>
+          <Typography color="textPrimary">Breadcrumb</Typography>
         </Breadcrumbs>
       </Paper>
     </div>
   );
 }
 
-SimpleBreadcrumbs.propTypes = {
+CustomSeparator.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(styles)(SimpleBreadcrumbs);
+export default withStyles(styles)(CustomSeparator);
