+++ docs/src/pages/demos/breadcrumbs/breadcrumbs.md
@@ -9,21 +9,21 @@ components: Breadcrumbs, Link, Typography
 
 ## Simple breadcrumbs
 
-{{"demo": "pages/lab/breadcrumbs/SimpleBreadcrumbs.js"}}
+{{"demo": "pages/demos/breadcrumbs/SimpleBreadcrumbs.js"}}
 
 ## Custom separator
 
 In the following examples, we are using two string separators, and an SVG icon.
 
-{{"demo": "pages/lab/breadcrumbs/CustomSeparator.js"}}
+{{"demo": "pages/demos/breadcrumbs/CustomSeparator.js"}}
 
 ## Breadcrumbs with icons
 
-{{"demo": "pages/lab/breadcrumbs/IconBreadcrumbs.js"}}
+{{"demo": "pages/demos/breadcrumbs/IconBreadcrumbs.js"}}
 
 ## Collapsed breadcrumbs
 
-{{"demo": "pages/lab/breadcrumbs/CollapsedBreadcrumbs.js"}}
+{{"demo": "pages/demos/breadcrumbs/CollapsedBreadcrumbs.js"}}
 
 ## Customized breadcrumbs
 
@@ -31,7 +31,7 @@ If you have been reading the [overrides documentation page](/customization/overr
 but you are not confident jumping in,
 here is one example of how you can change the breadcrumb link design.
 
-{{"demo": "pages/lab/breadcrumbs/CustomizedBreadcrumbs.js"}}
+{{"demo": "pages/demos/breadcrumbs/CustomizedBreadcrumbs.js"}}
 
 ## Accessibility
 
@@ -45,4 +45,4 @@ The accessibility of this component relies on:
 
 ## Integration with react-router
 
-{{"demo": "pages/lab/breadcrumbs/RouterBreadcrumbs.js"}}
+{{"demo": "pages/demos/breadcrumbs/RouterBreadcrumbs.js"}}
