+++ packages/material-ui-lab/src/index.d.ts
@@ -1,4 +1,3 @@
-export { default as Breadcrumbs } from './Breadcrumbs';
 export { default as Slider } from './Slider';
 export { default as SpeedDial } from './SpeedDial';
 export { default as SpeedDialAction } from './SpeedDialAction';
