+++ packages/material-ui/src/Breadcrumbs/BreadcrumbSeparator.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import { withStyles } from '@material-ui/core/styles';
+import withStyles from '../styles/withStyles';
 
 const styles = {
   root: {
