+++ pages/_app.js
@@ -289,9 +289,6 @@ const pages = [
         title: 'About The Lab',
       },
       {
-        pathname: '/lab/breadcrumbs',
-      },
-      {
         pathname: '/lab/slider',
       },
       {
