+++ pages/api/link.md
@@ -57,6 +57,6 @@ You can take advantage of this behavior to [target nested components](/guides/ap
 
 ## Demos
 
-- [Breadcrumbs](/lab/breadcrumbs/)
+- [Breadcrumbs](/demos/breadcrumbs/)
 - [Links](/style/links/)
