+++ pages/demos/breadcrumbs.js
@@ -3,13 +3,13 @@ import 'docs/src/modules/components/bootstrap';
 import React from 'react';
 import MarkdownDocs from 'docs/src/modules/components/MarkdownDocs';
 
-const req = require.context('docs/src/pages/lab/breadcrumbs', false, /\.md|\.js$/);
+const req = require.context('docs/src/pages/demos/breadcrumbs', false, /\.md|\.js$/);
 const reqSource = require.context(
-  '!raw-loader!../../docs/src/pages/lab/breadcrumbs',
+  '!raw-loader!../../docs/src/pages/demos/breadcrumbs',
   false,
   /\.js$/,
 );
-const reqPrefix = 'pages/lab/breadcrumbs';
+const reqPrefix = 'pages/demos/breadcrumbs';
 
 function Page() {
   return <MarkdownDocs req={req} reqSource={reqSource} reqPrefix={reqPrefix} />;
